requirements.txt
matplotlib
wordcloud
treetaggerwrapper
python-dotenv
pandas  # Optional, for data handling
nltk    # Optional, for additional text processing